+++ browser/utils/makeFrontendIframe.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import store from './store';
-
+import store from '../store';
 
 const makeFrontendIframe = () => {
 	let state = store.getState();
